+++ tests.py
@@ -1,21 +1,16 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright (c) 2016 -- Lars Heuer - Semagia <http://www.semagia.com/>.
+# Copyright (c) 2016 - 2020 -- Lars Heuer - Semagia <http://www.semagia.com/>.
 # All rights reserved.
 #
 # License: BSD License
 #
 """\
 Tests against Segno PIL/Pillow.
-
-:author:       Lars Heuer (heuer[at]semagia.com)
-:organization: Semagia - http://www.semagia.com/
-:license:      BSD License
 """
 from __future__ import absolute_import
 import pytest
 import segno
-from segno import colors
 
 
 def test_pil():
@@ -29,14 +24,14 @@ def test_pil():
 
 def test_pil_color():
     qr = segno.make_qr('A')
-    img = qr.to_pil(color='green')
+    img = qr.to_pil(dark='green')
     assert img
     assert 'P' == img.mode
 
 
 def test_pil_color_black():
     qr = segno.make_qr('A')
-    img = qr.to_pil(color='#000')
+    img = qr.to_pil(dark='#000')
     assert img
     assert '1' == img.mode, 'Expected greyscale image'
     assert qr.symbol_size() == img.size
@@ -44,23 +39,15 @@ def test_pil_color_black():
 
 def test_pil_greyscale_mirrored():
     qr = segno.make_qr('A')
-    img = qr.to_pil(color='#fff', background='#000')
+    img = qr.to_pil(dark='#fff', light='#000')
     assert img
     assert '1' == img.mode, 'Expected greyscale image'
     assert qr.symbol_size() == img.size
 
 
-def test_pil_greyscale_but_force_palette():
-    qr = segno.make_qr('A')
-    img = qr.to_pil(color='#000', mode='P')
-    assert img
-    assert 'P' == img.mode, 'Expected indexed-color image'
-    assert qr.symbol_size() == img.size
-
-
 def test_pil_pal_background():
     qr = segno.make('Hello')
-    img = qr.to_pil(color='white', background='green')
+    img = qr.to_pil(dark='white', light='green')
     assert img
     assert 'P' == img.mode
     assert qr.symbol_size() == img.size
@@ -92,90 +79,47 @@ def test_pil_border():
 
 def test_pil_transparent():
     qr = segno.make_qr('A')
-    color = 'black'
-    color_rgb = colors.color_to_rgb(color)
-    color_inverse = colors.invert_color(color_rgb)
-    img = qr.to_pil(background=None)  # color = black is implicit
+    img = qr.to_pil(light=None)  # color = black is implicit
     assert img
-    assert 'P' == img.mode, 'Expected indexed-color img'
+    assert '1' == img.mode
     assert qr.symbol_size() == img.size
-    palette = img.getpalette()
-    assert color_inverse == tuple(palette[:3])  # 1st entry: background
-    assert color_rgb == tuple(palette[3:6])     # 2nd entry: stroke color
+    assert 'transparency' in img.info
 
 
 def test_pil_color_black_tranparent():
     qr = segno.make_qr('A')
-    color = 'black'
-    color_rgb = colors.color_to_rgb(color)
-    color_inverse = colors.invert_color(color_rgb)
-    img = qr.to_pil(color=color, background=None)
+    img = qr.to_pil(dark='black', light=None)
     assert img
-    assert 'P' == img.mode, 'Expected indexed-color img'
+    assert '1' == img.mode
     assert qr.symbol_size() == img.size
-    palette = img.getpalette()
-    assert color_inverse == tuple(palette[:3])  # 1st entry: background
-    assert color_rgb == tuple(palette[3:6])     # 2nd entry: stroke color
+    assert 'transparency' in img.info
 
 
-def test_pil_color_white_tranparent():
+def test_pil_color_white_transparent():
     qr = segno.make_qr('A')
-    color = '#fff'
-    color_rgb = colors.color_to_rgb(color)
-    color_inverse = colors.invert_color(color_rgb)
-    img = qr.to_pil(color=color, background=None)
+    img = qr.to_pil(dark='#fff', light=None)
     assert img
-    assert 'P' == img.mode, 'Expected indexed-color img'
+    assert '1' == img.mode
     assert qr.symbol_size() == img.size
-    palette = img.getpalette()
-    assert color_inverse == tuple(palette[:3])  # 1st entry: background
-    assert color_rgb == tuple(palette[3:6])     # 2nd entry: stroke color
+    assert 'transparency' in img.info
 
 
-def test_pil_color_other_tranparent():
+def test_pil_color_other_transparent():
     qr = segno.make_qr('A')
-    color = 'green'
-    color_rgb = colors.color_to_rgb(color)
-    color_inverse = colors.invert_color(color_rgb)
-    img = qr.to_pil(color=color, background=None)
+    img = qr.to_pil(dark='green', light=None)
     assert img
     assert 'P' == img.mode, 'Expected indexed-color image'
     assert qr.symbol_size() == img.size
-    palette = img.getpalette()
-    assert color_inverse == tuple(palette[:3])  # 1st entry: background
-    assert color_rgb == tuple(palette[3:6])     # 2nd entry: stroke color
+    assert 'transparency' in img.info
 
 
-def test_pil_rgba_autodetect():
+def test_pil_palette_autodetect():
     qr = segno.make('Segno')
-    color = '#00fc'
-    color_rgba = colors.color_to_rgba(color, alpha_float=False)
-    color_inverse = colors.invert_color(color_rgba[:3]) + (0,)
-    img = qr.to_pil(color=color, background=None, border=0)
+    img = qr.to_pil(dark='#00fc', light=None, border=0)
     assert img
-    assert 'RGBA' == img.mode, 'Expected RGBA image'
+    assert 'P' == img.mode, 'Expected indexed-color image'
     assert qr.symbol_size(border=0) == img.size
-    assert color_rgba == img.getpixel((0, 0))
-    assert color_inverse == img.getpixel((1, 1))
-
-
-def test_pil_mode_illegal():
-    qr = segno.make('Segno')
-    with pytest.raises(ValueError):
-        qr.to_pil(scale=10, mode='U')
-
-
-def test_pil_mode_illegal2():
-    qr = segno.make('Segno')
-    with pytest.raises(ValueError):
-        qr.to_pil(scale=10, mode='RGB')
-
-
-def test_pil_mode_force_rgba():
-    qr = segno.make('Segno')
-    img = qr.to_pil(mode='RGBA')
-    assert 'RGBA' == img.mode
-    assert qr.symbol_size() == img.size
+    assert 'transparency' in img.info
 
 
 if __name__ == '__main__':